Remove useIP, detect IP with axios in helpers.
